// src/store/types/V2CommentPayload.ts
/**
 * Represents a reply to a comment thread, or the original comment of the thread.
 * A thread with no replies will have just the 1 original comment.
 * A thread with replies will have 2 or more comments.
 */
export type V2CommentPayload = {
  /* eslint-disable camelcase */
  /**
   * User id of the thread author
   */
  author_id: number
  /**
   * Actual comment body
   */
  body: string
  /** temporary field indicating if the comment is generated by the system */
  created_by_system?: boolean
  /**
   * UUID of the parent comment thread
   */
  comment_thread_id: string
  /**
   * UUID of the comment itself
   */
  id: string
  /**
   * ISO timestamp of when the comment was created.
   * The comments should be sorted by this to be listed in correct order.
   */
  inserted_at: string
  /**
   * ISO timestamp of when the comment was last edited.
   * Same as `inserted_at` if it was not edited.
   */
  updated_at: string
  /* eslint-enable camelcase */
}